pyproject.toml
[tool.poetry]
name = "readconnect-back"
version = "0.1.0"
description = ""
authors = ["Yeison Fernandez"]
readme = "README.md"
packages = [{ include = "readconnect", from = "src" }]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.24.0" }
python-json-logger = "^2.0.7"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
nanoid = "^2.0.0"
mysql-connector-python = "^8.2.0"
pydantic = {extras = ["email"], version = "^2.4.2"}
asyncpg = "^0.29.0"
greenlet = "^3.0.1"
pymonad = "^2.4.0"
install = "^1.3.5"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
black = "^23.10.1"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
start = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
